pyproject.toml
[project]
name = "integrando-python-com-open-ai-gpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
]
